document layer better & break out
